+++ spec/rubocop/cop/rspec/contain_exactly_spec.rb
@@ -1,10 +1,8 @@
 # frozen_string_literal: true
 
 RSpec.describe RuboCop::Cop::RSpec::ContainExactly do
-  it 'flags `contain_exactly` with only splat arguments' do
+  it 'flags `contain_exactly` with a single splat argument' do
     expect_offense(<<~RUBY)
-      it { is_expected.to contain_exactly(*array1, *array2) }
-                          ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ Prefer `match_array` when matching array values.
       it { is_expected.to contain_exactly(*[1,2,3]) }
                           ^^^^^^^^^^^^^^^^^^^^^^^^^ Prefer `match_array` when matching array values.
       it { is_expected.to contain_exactly(*a.merge(b)) }
@@ -14,7 +12,6 @@ RSpec.describe RuboCop::Cop::RSpec::ContainExactly do
     RUBY
 
     expect_correction(<<~RUBY)
-      it { is_expected.to match_array(array1 + array2) }
       it { is_expected.to match_array([1,2,3]) }
       it { is_expected.to match_array(a.merge(b)) }
       it { is_expected.to match_array((a + b)) }
@@ -42,6 +39,13 @@ RSpec.describe RuboCop::Cop::RSpec::ContainExactly do
     expect_no_offenses(<<~RUBY)
       it { is_expected.to contain_exactly(content, *array) }
       it { is_expected.to contain_exactly(*array, content) }
+    RUBY
+  end
+
+  it 'does not flag `contain_exactly` with multiple splat arguments' do
+    expect_no_offenses(<<~RUBY)
+      it { is_expected.to contain_exactly(*array1, *array2) }
+      it { is_expected.to contain_exactly(*0..9, *100..109, *200..209) }
     RUBY
   end